Thêm đăng nhập và footer

+++ WebSach/WebSach/Areas/WebAdmin/Controllers/AdminUsersController.cs
@@ -1,13 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using WebSach.Models;
+using System.Security.Cryptography;
+using System.Text;
+using System.Web.UI.WebControls;
+using System.Windows;
 
 namespace WebSach.Areas.WebAdmin.Controllers
 {
@@ -18,12 +20,16 @@ namespace WebSach.Areas.WebAdmin.Controllers
         // GET: WebAdmin/AdminUsers
         public async Task<ActionResult> Index()
         {
+            if (Session["Admin"] == null)
+                return View("Login");
             return View(await db.User.ToListAsync());
         }
 
         // GET: WebAdmin/AdminUsers/Details/5
         public async Task<ActionResult> Details(string id)
         {
+            if (Session["Admin"] == null)
+                return View("Login");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +45,8 @@ namespace WebSach.Areas.WebAdmin.Controllers
         // GET: WebAdmin/AdminUsers/Create
         public ActionResult Create()
         {
+            if (Session["Admin"] == null)
+                return View("Login");
             return View();
         }
 
@@ -62,6 +70,8 @@ namespace WebSach.Areas.WebAdmin.Controllers
         // GET: WebAdmin/AdminUsers/Edit/5
         public async Task<ActionResult> Edit(string id)
         {
+            if (Session["Admin"] == null)
+                return View("Login");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -93,6 +103,8 @@ namespace WebSach.Areas.WebAdmin.Controllers
         // GET: WebAdmin/AdminUsers/Delete/5
         public async Task<ActionResult> Delete(string id)
         {
+            if (Session["Admin"] == null)
+                return View("Login", "AdminUsers");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -111,8 +123,16 @@ namespace WebSach.Areas.WebAdmin.Controllers
         public async Task<ActionResult> DeleteConfirmed(string id)
         {
             User user = await db.User.FindAsync(id);
-            db.User.Remove(user);
-            await db.SaveChangesAsync();
+            var find = db.Books.Where(f => f.User_Name == user.User_Name);
+            if (find != null)
+            {
+                MessageBox.Show("Người dùng có liên kết với những dữ liệu khác. Hãy xóa dữ liệu của người dùng trước", "Alert");
+            }
+            else
+            {
+                db.User.Remove(user);
+                await db.SaveChangesAsync();
+            }
             return RedirectToAction("Index");
         }
 
@@ -123,6 +143,66 @@ namespace WebSach.Areas.WebAdmin.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        // LOGIN
+        public ActionResult Login()
+        {
+            return View();
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Login(User _user)
+        {
+            if (ModelState.IsValid)
+            {
+                var f_password = _user.
+                    Password;
+                var data = db.User.Where(s => s.User_Name.Equals(_user.User_Name) && s.Password.Equals(f_password) && s.Permission_Id == true).ToList();
+                if (data.Count() > 0)
+                {
+                    //add session
+                    Session["Admin"] = data.FirstOrDefault().User_Name;
+                    data.FirstOrDefault().Last_Login = DateTime.Now;
+                    return RedirectToAction("Index", "HomeAdmin");
+                }
+                else
+                {
+                    ViewBag.error = "Invalid username or password!";
+                    return View("Login");
+                }
+            }
+            else
+            {
+                ViewBag.error = "Cannot login";
+                return View();
+
+            }
+        }
+        //Logout
+        public ActionResult Logout()
+        {
+            Session.Clear();//remove session
+            return RedirectToAction("Index", "Home");
+        }
+
+
+
+        //create a string MD5
+        public static string GetMD5(string str)
+        {
+            MD5 md5 = new MD5CryptoServiceProvider();
+            byte[] fromData = Encoding.UTF8.GetBytes(str);
+            byte[] targetData = md5.ComputeHash(fromData);
+            string byte2String = null;
+
+            for (int i = 0; i < targetData.Length; i++)
+            {
+                byte2String += targetData[i].ToString("x2");
+
+            }
+            return byte2String;
         }
     }
 }
